move most CounterGather tests over to index protocol tests
